feat: calendar restructure
