+++ frontend/src/components/ui/calendar/index.ts
@@ -1,0 +1,2 @@
+export { Calendar } from './calendar';
+export type { CalendarProps } from './calendar';
